tweeking ploting

+++ plot.py
@@ -327,7 +327,7 @@ f1.close
 
 for j in range(0,len(QI)):
   hold=0
-  for i in range(len(EiVe)*0.6,len(EiVe)):
+  for i in range(int(len(EiVe)*0.6),len(EiVe)):
       hold=hold+(EiVe[int(i)][QI[int(j)]]*EiVe[int(i)][QI[int(j)]].conjugate()).real
   while hold>0.05:
     QI[int(j)]=QI[int(j)]+1
